fix : change to Double

+++ src/src/main/kotlin/main.kt
@@ -1,10 +1,10 @@
 fun main(args: Array<String>) {
 
     print("Введите первое значение: ")
-    var firstNumber = readLine()!!.toInt()
+    var firstNumber = readLine()!!.toDouble()
 
     print("Введите второе значение: ")
-    var secondNumber = readLine()!!.toInt()
+    var secondNumber = readLine()!!.toDouble()
 
     print("Введите операцию: ")
     var operation = readLine()
